hub/requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
graphql-core==3.2.3
graphql-server==3.0.0b5
graphene==3.2.1
PyYAML==6.0
pymongo==4.3.3
motor==3.1.1
simplejson
pytz
jinja2
